refactor(playground-ui): move autoform from ds to lib/form
