+++ .changeset/move-autoform-to-lib-form.md
@@ -1,0 +1,5 @@
+---
+"@mastra/playground-ui": patch
+---
+
+Move AutoForm from ds/components to lib/form for better organization
